Refactor game logic and UI updates for seed planting and harvesting - Removed the old UI code from js/ui.js and replaced it with a modular structure in uiUpdates.js. - Introduced seed data management in seedData.js, including detailed properties for each seed type. - Enhanced game logic in game.js to handle seed buying, planting, harvesting, and selling with improved weight and value calculations. - Updated game state management in gameState.js to include tools and inventory initialization. - Added event listeners in eventListeners.js for user interactions with the UI. - Improved the main game loop in main.js to ensure smooth gameplay and state updates. - Implemented a new tools display feature to show available tools in the inventory.
