+++ index.html
@@ -63,9 +63,6 @@
         </div>
     </div>
 
-    <script src="js/data.js"></script>
-    <script src="js/game.js"></script>
-    <script src="js/ui.js"></script>
-    <script src="js/logic.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
